requirements.txt
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.3.2
billiard==3.3.0.20
celery==3.1.18
Django==1.8.1
django-adminplus==0.3
django-annoying==0.8.1
django-braces==1.8.0
django-configurations==0.8
django-countries==3.3
django-model-utils==2.2
django-pipeline==1.5.0
django-webtest==1.7.8
django-widget-tweaks==1.3
djangorestframework==3.1.2
ecdsa==0.13
Fabric==1.8.1
futures==3.0.1
geopy==1.11.0
kombu==3.0.26
lockfile==0.10.2
mangopaysdk==0.311
model-mommy==1.2.4
oauthlib==0.7.2
paramiko==1.15.2
Pillow==2.8.1
psycopg2==2.6
pycrypto==2.6.1
pyfcm==1.0.5
python-memcached==1.57
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sorl-thumbnail==12.2
suds-jurko==0.6
uWSGI==2.0.11.1
waitress==0.8.9
WebOb==1.4.1
WebTest==2.0.18
wheel==0.24.0
